# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sorolens API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Sorolens indexes and surfaces data from Soroban smart contracts: events,
    invocations, storage, and per-contract analytics. This spec is the single
    source of truth for the auto-generated Go client (packages/go-client) and
    matches the live routes in apps/api/internal/router/router.go.

    Two namespaces are documented: the original `/api/v1` surface and the
    consistent `/api/v2` surface introduced in issue #144. v2 is additive; v1 is
    unchanged. See docs/api-v2.md for the field-by-field v1 -> v2 mapping.

    Authentication: most read endpoints accept an optional API key via
    `Authorization: Bearer <token>` or `X-API-Key: <token>`. API-key management
    additionally requires an admin identity via role headers
    (`X-User-ID`, `X-GitHub-Login`, `X-User-Sub`, or `X-GitHub-ID`).
  contact:
    name: Sorolens
servers:
  - url: https://api.sorolens.xyz
    description: Production
...
paths:
...
  /api/v1/contracts:
    get:
      tags: [Contracts]
      summary: List tracked contracts
      operationId: listContracts
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/NetworkParam"
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by contract status (pending, active, backfilling, error, paused).
      responses:
        "200":
          description: Page of contracts
          content:
            application/json:
              schema:
                type: object
                required: [contracts, next_cursor]
                properties:
                  contracts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Contract"
                  next_cursor:
                    type: string
                    description: Opaque cursor for the next page; empty when no more pages.
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Contracts]
      summary: Register a contract for tracking
      operationId: registerContract
      security:
        - ApiKeyAuth: [write:contracts]
        - RoleAuth: [contributor]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, network]
              properties:
                id:
                  type: string
                  pattern: "^C[A-Za-z0-9]{55}$"
                  description: Stellar contract ID (56 chars, starts with C).
                network:
                  type: string
                  enum: [testnet, mainnet, futurenet, standalone]
                label:
                  type: string
                  description: Optional human-friendly label.
      responses:
        "201":
          description: Contract registered (status pending until backfill completes).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/InvalidInput"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/contracts/{id}:
    get:
      tags: [Contracts]
      summary: Get one contract
      operationId: getContract
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
      responses:
        "200":
          description: The contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/contracts/{id}/stats:
    get:
      tags: [Stats]
      summary: Per-contract statistics
      operationId: getContractStats
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - $ref: "#/components/parameters/ContractID"
        - name: window
          in: query
          required: false
          schema:
            type: string
          description: Aggregation window. 7d and 30d map to 7/30 days; anything else is 24 hours.
      responses:
        "200":
          description: Contract statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractStats"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/events/recent:
    get:
      tags: [Live]
      summary: Newest events across every tracked contract
      operationId: recentEvents
      security:
        - ApiKeyAuth: [read:contracts]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Recent events, newest first
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/Event"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "Scoped API key. Send as `Authorization: Bearer <token>` (or `X-API-Key`)."
    RoleAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        Role-based auth. Resolve a user identity from one of
        X-User-ID, X-GitHub-Login, X-User-Sub, or X-GitHub-ID and enforce the
        minimum role declared next to this security scheme.
    IdentityHeaderAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Watchlist endpoints read the user ID from X-User-ID.
  parameters:
    ContractID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Contract ID (starts with C).
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque pagination cursor from a previous response.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
      description: Page size (default 50, clamped to 200).
    NetworkParam:
      name: network
      in: query
      required: false
      schema:
        type: string
        enum: [testnet, mainnet, futurenet, standalone]
      description: Network filter; empty means all networks.
...
  responses:
    Unauthorized:
      description: Missing or invalid credential or identity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ForbiddenRole:
      description: Identity present but role below the required minimum
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RoleError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidInput:
      description: Invalid query or body input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnsupportedMediaType:
      description: Content-Type is not application/json
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Global per-IP rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          oneOf:
            - type: object
              required: [code, message]
              properties:
                code:
                  type: string
                  enum: [NOT_FOUND, INVALID_INPUT, INTERNAL, RATE_LIMITED, UNSUPPORTED_MEDIA_TYPE]
                message:
                  type: string
                request_id:
                  type: string
            - type: object
              properties:
                error:
                  type: string
...
    RoleError:
      type: object
      required: [error, required, role]
      properties:
        error:
          type: string
        required:
          type: string
        role:
          type: string
    Contract:
      type: object
      required: [id, network, label, wasm_hash, created_at_ledger, backfill_complete_at, status, added_at]
      properties:
        id:
          type: string
        network:
          type: string
        label:
          type: string
        wasm_hash:
          type: string
        created_at_ledger:
          type: integer
          format: int64
        backfill_complete_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, active, backfilling, error, paused]
        added_at:
          type: string
          format: date-time
    Event:
      type: object
      required:
        - id
        - contract_id
        - network
        - ledger
        - ledger_closed_at
        - tx_hash
        - type
        - topic_xdr
        - value_xdr
        - topic_decoded
        - value_decoded
        - in_successful_call
      properties:
        id:
          type: string
        contract_id:
          type: string
        network:
          type: string
        ledger:
          type: integer
          format: int32
        ledger_closed_at:
          type: string
          format: date-time
        tx_hash:
          type: string
        type:
          type: string
        topic_xdr:
          type: array
          items:
            type: string
        value_xdr:
          type: string
        topic_decoded:
          type: array
          items: {}
        value_decoded: {}
        in_successful_call:
          type: boolean
...
    ContractStats:
      type: object
      required:
        - event_count
        - invocation_count
        - storage_count
        - last_synced_ledger
        - window_event_count
        - window_invocation_count
        - window_duration
      properties:
        event_count:
          type: integer
          format: int64
        invocation_count:
          type: integer
          format: int64
        storage_count:
          type: integer
          format: int64
        last_synced_ledger:
          type: integer
          format: int32
        window_event_count:
          type: integer
          format: int64
        window_invocation_count:
          type: integer
          format: int64
        window_duration:
          type: string
...